fix rpc count

+++ tcvectordb/src/main/java/com/tencent/tcvectordb/model/param/entity/BaseRes.java
@@ -6,7 +6,7 @@ public class BaseRes {
     protected String msg;
     protected String warning;
 
-    protected int count;
+    protected long count;
 
     protected String requestId;
 
@@ -19,7 +19,7 @@ public class BaseRes {
         this.warning = warning;
     }
 
-    public BaseRes(int code, String msg, String warning, int count) {
+    public BaseRes(int code, String msg, String warning, long count) {
         this.code = code;
         this.msg = msg;
         this.warning = warning;
@@ -46,7 +46,7 @@ public class BaseRes {
         return warning;
     }
 
-    public int getCount() {
+    public long getCount() {
         return count;
     }
